// simulation/web_platform/robot_benchmark/application/spa/src/shared/api/data-contracts.ts
export interface ProblemUser {
    id: number;
    is_completed?: boolean;
    /**
     * @min -2147483648
     * @max 2147483647
     */
    points?: number;
    robot_panel_port: number;
    vs_port: number;
    webots_stream_port: number;
    user: number;
    problem: number;
}

export interface TokenRefresh {
    access: string;
    refresh: string;
}

/** Сериализатор для модель соревнования */
export interface Tournament {
    id: number;
    users: User[];
    /** @maxLength 255 */
    name: string;
    /** @maxLength 5000 */
    description: string;
    /** @format date-time */
    date_start: Date;
    /** @format date-time */
    date_end: Date;
    problems: number[];
}

/** Сериализатор для модель many-to-many Соревнования-Пользователи */
export interface TournamentUser {
    id: number;
    is_completed?: boolean;
    /**
     * @min -2147483648
     * @max 2147483647
     */
    points?: number;
    user: number;
    tournament: number;
}

export interface User {
    id: number;
    /** @maxLength 128 */
    password: string;
    /**
     * Superuser status
     * Designates that this user has all permissions without explicitly assigning them.
     */
    is_superuser?: boolean;
    /**
     * Required. 150 characters or fewer. Letters, digits and @/./+/-/_ only.
     * @maxLength 150
     * @pattern ^[\w.@+-]+$
     */
    username: string;
    /** @maxLength 150 */
    first_name?: string;
    /** @maxLength 150 */
    last_name?: string;
    /**
     * Email address
     * @format email
     * @maxLength 254
     */
    email?: string;
}